Conflicts:
	liquidvirtue/views.py

+++ liquidvirtue/views.py
@@ -86,22 +86,29 @@ def get_next(request):
 
 def like(request, lv_video_id):
 	like_status = ''
-	if Like.objects.filter(user=request.user.id).filter(video=lv_video_id).exists():
-		v = Video.objects.get(pk=lv_video_id)
-		v.num_likes = v.num_likes - 1
-		if v.num_likes < 0:
-			v.num_likes = 0
-		v.save()
-		l = Like.objects.filter(user=request.user.id).filter(video=lv_video_id)
-		l.delete()
-		like_status = 'plus'
-	else:
-		v = Video.objects.get(pk=lv_video_id)
-		v.num_likes = v.num_likes + 1
-		v.save()
-		l = Like( video=v, user=UserProfile.objects.get(pk=request.user.id) )
-		l.save()
-		like_status = 'heart'
+	if request.user.is_authenticated:
+		userprofile = False
+		try:
+			userprofile = UserProfile.objects.get(pk=request.user.id)
+		except UserProfile.DoesNotExist:
+			pass
+		if userprofile:
+			if Like.objects.filter(user=request.user.id).filter(video=lv_video_id).exists():
+				v = Video.objects.get(pk=lv_video_id)
+				v.num_likes = v.num_likes - 1
+				if v.num_likes < 0:
+					v.num_likes = 0
+				v.save()
+				l = Like.objects.filter(user=request.user.id).filter(video=lv_video_id)
+				l.delete()
+				like_status = 'plus'
+			else:
+				v = Video.objects.get(pk=lv_video_id)
+				v.num_likes = v.num_likes + 1
+				v.save()
+				l = Like( video=v, user=UserProfile.objects.get(pk=request.user.id) )
+				l.save()
+				like_status = 'heart'
 	
 	response_data['like_status'] = like_status
 	return HttpResponse(json.dumps(response_data), mimetype="application/json")
@@ -183,39 +190,7 @@ def trackbox_newest(request, page_number):
 	class_names = []
 	
 	for video in videos:
-		year_delta = now.year - video.upload_date.year
-		month_delta = now.month - video.upload_date.month
-		day_delta = now.day - video.upload_date.day
-		hour_delta = now.hour - video.upload_date.hour
-		minute_delta = now.minute - video.upload_date.minute
-		second_delta = now.second - video.upload_date.second
-	
-		if year_delta > 1:
-			upload_date_text = 'Posted ' + str(year_delta) + ' years ago'
-		elif year_delta == 1:
-			upload_date_text = 'Posted ' + str(year_delta) + ' year ago'
-		elif month_delta > 1:
-			upload_date_text = 'Posted ' + str(month_delta) + ' months ago'
-		elif month_delta == 1:
-			upload_date_text = 'Posted ' + str(month_delta) + ' month ago'
-		elif day_delta > 1:
-			upload_date_text = 'Posted ' + str(day_delta) + ' days ago'
-		elif day_delta == 1:
-			upload_date_text = 'Posted ' + str(day_delta) + ' day ago'
-		elif hour_delta > 1:
-			upload_date_text = 'Posted ' + str(hour_delta) + ' hours ago'
-		elif hour_delta == 1:
-			upload_date_text = 'Posted ' + str(hour_delta) + ' hour ago'
-		elif minute_delta > 1:
-			upload_date_text = 'Posted ' + str(minute_delta) + ' minutes ago'
-		elif minute_delta == 1:
-			upload_date_text = 'Posted ' + str(minute_delta) + ' minute ago'
-		elif second_delta > 1:
-			upload_date_text = 'Posted ' + str(second_delta) + ' seconds ago'
-		elif second_delta == 1:
-			upload_date_text = 'Posted ' + str(second_delta) + ' second ago'
-		else:
-			upload_date_text = 'Posted now'
+		upload_date_text = get_upload_date_text( video.upload_date, now )
 		upload_date_texts.append(upload_date_text)
 		if Like.objects.filter(user=request.user.id).filter(video=video.id).exists():
 			class_names.append( 'heart' )
@@ -266,39 +241,7 @@ def trackbox_popular(request, page_number):
 	class_names = []
 	
 	for video in videos:
-		year_delta = now.year - video.upload_date.year
-		month_delta = now.month - video.upload_date.month
-		day_delta = now.day - video.upload_date.day
-		hour_delta = now.hour - video.upload_date.hour
-		minute_delta = now.minute - video.upload_date.minute
-		second_delta = now.second - video.upload_date.second
-	
-		if year_delta > 1:
-			upload_date_text = 'Posted ' + str(year_delta) + ' years ago'
-		elif year_delta == 1:
-			upload_date_text = 'Posted ' + str(year_delta) + ' year ago'
-		elif month_delta > 1:
-			upload_date_text = 'Posted ' + str(month_delta) + ' months ago'
-		elif month_delta == 1:
-			upload_date_text = 'Posted ' + str(month_delta) + ' month ago'
-		elif day_delta > 1:
-			upload_date_text = 'Posted ' + str(day_delta) + ' days ago'
-		elif day_delta == 1:
-			upload_date_text = 'Posted ' + str(day_delta) + ' day ago'
-		elif hour_delta > 1:
-			upload_date_text = 'Posted ' + str(hour_delta) + ' hours ago'
-		elif hour_delta == 1:
-			upload_date_text = 'Posted ' + str(hour_delta) + ' hour ago'
-		elif minute_delta > 1:
-			upload_date_text = 'Posted ' + str(minute_delta) + ' minutes ago'
-		elif minute_delta == 1:
-			upload_date_text = 'Posted ' + str(minute_delta) + ' minute ago'
-		elif second_delta > 1:
-			upload_date_text = 'Posted ' + str(second_delta) + ' seconds ago'
-		elif second_delta == 1:
-			upload_date_text = 'Posted ' + str(second_delta) + ' second ago'
-		else:
-			upload_date_text = 'Posted now'
+		upload_date_text = get_upload_date_text( video.upload_date, now )
 		upload_date_texts.append(upload_date_text)
 		if Like.objects.filter(user=request.user.id).filter(video=video.id).exists():
 			class_names.append( 'heart' )
@@ -325,46 +268,14 @@ def trackbox_my_library(request, page_number):
 	videos = Video.objects.filter(like__user__id__exact=request.user.id).order_by('-upload_time')[(page_number-1)*17:(page_number*17)]
 	#l = Like.objects.filter(user=request.user.id)
 	#videos = l.video_set.order_by('-upload_time')[(page_number-1)*17:(page_number*17)]
-	
+
 	now = datetime.now()
-	
+
 	upload_date_texts = []
 	class_names = []
 	
 	for video in videos:
-		year_delta = now.year - video.upload_date.year
-		month_delta = now.month - video.upload_date.month
-		day_delta = now.day - video.upload_date.day
-		hour_delta = now.hour - video.upload_date.hour
-		minute_delta = now.minute - video.upload_date.minute
-		second_delta = now.second - video.upload_date.second
-	
-		if year_delta > 1:
-			upload_date_text = 'Posted ' + str(year_delta) + ' years ago'
-		elif year_delta == 1:
-			upload_date_text = 'Posted ' + str(year_delta) + ' year ago'
-		elif month_delta > 1:
-			upload_date_text = 'Posted ' + str(month_delta) + ' months ago'
-		elif month_delta == 1:
-			upload_date_text = 'Posted ' + str(month_delta) + ' month ago'
-		elif day_delta > 1:
-			upload_date_text = 'Posted ' + str(day_delta) + ' days ago'
-		elif day_delta == 1:
-			upload_date_text = 'Posted ' + str(day_delta) + ' day ago'
-		elif hour_delta > 1:
-			upload_date_text = 'Posted ' + str(hour_delta) + ' hours ago'
-		elif hour_delta == 1:
-			upload_date_text = 'Posted ' + str(hour_delta) + ' hour ago'
-		elif minute_delta > 1:
-			upload_date_text = 'Posted ' + str(minute_delta) + ' minutes ago'
-		elif minute_delta == 1:
-			upload_date_text = 'Posted ' + str(minute_delta) + ' minute ago'
-		elif second_delta > 1:
-			upload_date_text = 'Posted ' + str(second_delta) + ' seconds ago'
-		elif second_delta == 1:
-			upload_date_text = 'Posted ' + str(second_delta) + ' second ago'
-		else:
-			upload_date_text = 'Posted now'
+		upload_date_text = get_upload_date_text( video.upload_date, now )
 		upload_date_texts.append(upload_date_text)
 		if Like.objects.filter(user=request.user.id).filter(video=video.id).exists():
 			class_names.append( 'heart' )
@@ -391,44 +302,12 @@ def trackbox_search(request, page_number):
 	videos = Video.objects.filter(title__icontains=search).order_by('-upload_time')[(page_number-1)*17:(page_number*17)]
 
 	now = datetime.now()
-	
+
 	upload_date_texts = []
 	class_names = []
 	
 	for video in videos:
-		year_delta = now.year - video.upload_date.year
-		month_delta = now.month - video.upload_date.month
-		day_delta = now.day - video.upload_date.day
-		hour_delta = now.hour - video.upload_date.hour
-		minute_delta = now.minute - video.upload_date.minute
-		second_delta = now.second - video.upload_date.second
-	
-		if year_delta > 1:
-			upload_date_text = 'Posted ' + str(year_delta) + ' years ago'
-		elif year_delta == 1:
-			upload_date_text = 'Posted ' + str(year_delta) + ' year ago'
-		elif month_delta > 1:
-			upload_date_text = 'Posted ' + str(month_delta) + ' months ago'
-		elif month_delta == 1:
-			upload_date_text = 'Posted ' + str(month_delta) + ' month ago'
-		elif day_delta > 1:
-			upload_date_text = 'Posted ' + str(day_delta) + ' days ago'
-		elif day_delta == 1:
-			upload_date_text = 'Posted ' + str(day_delta) + ' day ago'
-		elif hour_delta > 1:
-			upload_date_text = 'Posted ' + str(hour_delta) + ' hours ago'
-		elif hour_delta == 1:
-			upload_date_text = 'Posted ' + str(hour_delta) + ' hour ago'
-		elif minute_delta > 1:
-			upload_date_text = 'Posted ' + str(minute_delta) + ' minutes ago'
-		elif minute_delta == 1:
-			upload_date_text = 'Posted ' + str(minute_delta) + ' minute ago'
-		elif second_delta > 1:
-			upload_date_text = 'Posted ' + str(second_delta) + ' seconds ago'
-		elif second_delta == 1:
-			upload_date_text = 'Posted ' + str(second_delta) + ' second ago'
-		else:
-			upload_date_text = 'Posted now'
+		upload_date_text = get_upload_date_text( video.upload_date, now )
 		upload_date_texts.append(upload_date_text)
 		if Like.objects.filter(user=request.user.id).filter(video=video.id).exists():
 			class_names.append( 'heart' )
@@ -447,17 +326,40 @@ def trackinfo(request, lv_video_id):
 	title = video.title
 	channel_name = video.channel_name
 	watch_page_url = video.watch_page_url
-	
+
+	now = datetime.now()
+
+	upload_date_text = get_upload_date_text( video.upload_date, now )
+
+	likes = Like.objects.filter(user=request.user.id, video_id=lv_video_id)
+	class_name = ''
+	alt_text = ''
+	if likes:
+		class_name = 'heart'
+		alt_text = 'Remove this track from your playlist'
+	else:
+		class_name = 'plus'
+		alt_text = 'Add this track to your playlist'
+		
+	response_data['class_name'] = class_name
+	response_data['alt_text'] = alt_text
+	response_data['title'] = title
+	response_data['channel_name'] = channel_name
+	response_data['upload_date_text'] = upload_date_text
+	response_data['lv_video_id'] = lv_video_id
+	response_data['watch_page_url'] = watch_page_url
+	return HttpResponse(json.dumps(response_data), mimetype="application/json")
+	#return render_to_response('trackinfo.html', {'class_name': class_name, 'alt_text': alt_text, 'title': title, 'channel_name': channel_name, 'upload_date_text': upload_date_text, 'lv_video_id': lv_video_id, 'watch_page_url': watch_page_url})
+
+def get_upload_date_text(upload_date, now):
 	upload_date_text = ''
 	
-	now = datetime.now()
-	
-	year_delta = now.year - video.upload_date.year
-	month_delta = now.month - video.upload_date.month
-	day_delta = now.day - video.upload_date.day
-	hour_delta = now.hour - video.upload_date.hour
-	minute_delta = now.minute - video.upload_date.minute
-	second_delta = now.second - video.upload_date.second
+	year_delta = now.year - upload_date.year
+	month_delta = now.month - upload_date.month
+	day_delta = now.day - upload_date.day
+	hour_delta = now.hour + 8 - upload_date.hour
+	minute_delta = now.minute - upload_date.minute
+	second_delta = now.second - upload_date.second
 
 	if year_delta > 1:
 		upload_date_text = 'Posted ' + str(year_delta) + ' years ago'
@@ -486,22 +388,4 @@ def trackinfo(request, lv_video_id):
 	else:
 		upload_date_text = 'Posted now'
 
-	likes = Like.objects.filter(user=request.user.id, video_id=lv_video_id)
-	class_name = ''
-	alt_text = ''
-	if likes:
-		class_name = 'heart'
-		alt_text = 'Remove this track from your playlist'
-	else:
-		class_name = 'plus'
-		alt_text = 'Add this track to your playlist'
-		
-	response_data['class_name'] = class_name
-	response_data['alt_text'] = alt_text
-	response_data['title'] = title
-	response_data['channel_name'] = channel_name
-	response_data['upload_date_text'] = upload_date_text
-	response_data['lv_video_id'] = lv_video_id
-	response_data['watch_page_url'] = watch_page_url
-	return HttpResponse(json.dumps(response_data), mimetype="application/json")
-	#return render_to_response('trackinfo.html', {'class_name': class_name, 'alt_text': alt_text, 'title': title, 'channel_name': channel_name, 'upload_date_text': upload_date_text, 'lv_video_id': lv_video_id, 'watch_page_url': watch_page_url})
+	return upload_date_text
